Changed core version, commented javadoc gen
Core version changed to the new one with slightly fewer hard-coded
stuff. Old version was overwritten in the process for some ungodly
reason with github. Javadoc generation is to remain commented out
until we create a deploy phase structure for this project, at
which point javadoc generation will be placed there. Generating
javadoc for testing phases (including package) is cumbersome.